Now the Path are matched by the order of definition
Now the filters and routes are matched based on the order of their
definition. Declaring two equivalent paths will cause the latter one
never be called (, and for two exact same paths, the later one will not
be loaded by the controller).
Detail:
The implementing Path in the router map now based on a new class,
TimedPath, which compares to other TimedPath based on the time of
creation
The implementing Route/Filter Map type changed from HashMap TreeMap
Reformatted the code for putting in routes and filters to support the
TimedPath